try_docker/imgtogcode/requirements.txt
Package         Version
--------------- ----------
click           8.1.3
colorama        0.4.6
contourpy       1.0.6
cycler          0.11.0
Flask           2.2.2
fonttools       4.38.0
imageio         2.23.0
itsdangerous    2.1.2
Jinja2          3.1.2
kiwisolver      1.4.4
MarkupSafe      2.1.1
matplotlib      3.6.2
networkx        2.8.8
numpy           1.24.0
opencv-python   4.6.0.66
packaging       22.0
Pillow          9.3.0
pip             22.3.1
pyparsing       3.0.9
python-dateutil 2.8.2
PyWavelets      1.4.1
rdp             0.8
scikit-image    0.19.3
scipy           1.9.3
setuptools      63.2.0
six             1.16.0
tifffile        2022.10.10
Werkzeug        2.2.2